refactor/#4: screen 별 navigation 패키지 생성
