+++ app/src/main/java/com/wearerommies/roomie/presentation/ui/home/navigation/HomeNavigation.kt
@@ -1,4 +1,4 @@
-package com.wearerommies.roomie.presentation.ui.mypage
+package com.wearerommies.roomie.presentation.ui.home.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
@@ -6,17 +6,18 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.wearerommies.roomie.presentation.navigator.route.MainTabRoute
+import com.wearerommies.roomie.presentation.ui.home.HomeRoute
 
-fun NavController.navigateMy(navOptions: NavOptions) {
+fun NavController.navigateHome(navOptions: NavOptions) {
     navigate(
-        route = MainTabRoute.My,
+        route = MainTabRoute.Home,
         navOptions = navOptions
     )
 }
 
-fun NavGraphBuilder.myNavGraph() {
-    composable<MainTabRoute.My> {
-        MyRoute(
+fun NavGraphBuilder.homeNavGraph() {
+    composable<MainTabRoute.Home>{
+        HomeRoute(
             paddingValues = PaddingValues(),
             navigateUp = {}
         )
