+++ app/src/main/java/com/wearerommies/roomie/presentation/ui/mypage/navigation/MyNavigation.kt
@@ -1,4 +1,4 @@
-package com.wearerommies.roomie.presentation.ui.map
+package com.wearerommies.roomie.presentation.ui.mypage.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
@@ -6,17 +6,18 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.wearerommies.roomie.presentation.navigator.route.MainTabRoute
+import com.wearerommies.roomie.presentation.ui.mypage.MyRoute
 
-fun NavController.navigateMap(navOptions: NavOptions) {
+fun NavController.navigateMy(navOptions: NavOptions) {
     navigate(
-        route = MainTabRoute.Map,
+        route = MainTabRoute.My,
         navOptions = navOptions
     )
 }
 
-fun NavGraphBuilder.mapNavGraph() {
-    composable<MainTabRoute.Map> {
-        MapRoute(
+fun NavGraphBuilder.myNavGraph() {
+    composable<MainTabRoute.My> {
+        MyRoute(
             paddingValues = PaddingValues(),
             navigateUp = {}
         )
